+++ src/components/lottie/checklottie.jsx
