+++ src/components/main/Mainsecond.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import Creditlottie from './Creditlottie.jsx'
+import Creditlottie from '../lottie/Creditlottie.jsx'
 import slideimage from '../../assets/images/mainsecond.png'
 
 function Mainsecond() {
